+++ src/app/core/entity-list/duplicate-records/duplicate-records.service.ts
@@ -1,13 +1,12 @@
-import { Injectable } from '@angular/core';
-import { EntityMapperService } from '../entity/entity-mapper/entity-mapper.service'; 
-import { EntityRegistry } from '../entity/database-entity.decorator';
-import { EntitySchemaService } from '../entity/schema/entity-schema.service';
-import { Entity } from '../entity/model/entity';
+import { Injectable } from "@angular/core";
+import { EntityMapperService } from "../../entity/entity-mapper/entity-mapper.service";
+import { EntityRegistry } from "../../entity/database-entity.decorator";
+import { EntitySchemaService } from "../../entity/schema/entity-schema.service";
+import { Entity } from "../../entity/model/entity";
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: "root",
 })
-
 export class DuplicateRecordService {
   get: jasmine.Spy<jasmine.Func>;
   constructor(
@@ -16,7 +15,7 @@ export class DuplicateRecordService {
     private entityService: EntitySchemaService,
   ) {}
 
-  async duplicateRecord(sourceData: Entity[])  {
+  async duplicateRecord(sourceData: Entity[]) {
     const duplicateData = this.clone(sourceData);
     return await this.entitymapperservice.saveAll(duplicateData);
   }
@@ -24,21 +23,26 @@ export class DuplicateRecordService {
   clone(sourceData: Entity[]): any {
     const duplicateData = [];
 
-    sourceData.map((item: Entity)=> {
+    sourceData.map((item: Entity) => {
       const entityConstructor = item.getConstructor();
-      const keys = [...entityConstructor.schema.keys()].filter(key => key !== '_id' && key !== '_rev');
+      const keys = [...entityConstructor.schema.keys()].filter(
+        (key) => key !== "_id" && key !== "_rev",
+      );
       const dbEntity = this.entityService.transformEntityToDatabaseFormat(item);
-      const entityformat = this.entityService.transformDatabaseToEntityFormat(dbEntity, entityConstructor.schema);
+      const entityformat = this.entityService.transformDatabaseToEntityFormat(
+        dbEntity,
+        entityConstructor.schema,
+      );
       const entity = new entityConstructor();
       const attribute = entityConstructor.toStringAttributes[0];
       for (const key of keys) {
         if (attribute === key) {
           entityformat[key] = `Copy of ${entityformat[key]}`;
         }
-       entity[key] = entityformat[key];
+        entity[key] = entityformat[key];
       }
       duplicateData.push(entity);
-    })
+    });
     return duplicateData;
   }
 }
